modified package structure - pt2
